gui_interence_app/requirements.txt
numpy
tensorflow
keras
pillow
django